+++ types/index.d.ts
@@ -10,3 +10,4 @@
 /// <reference path="./runtime-process.d.ts" />
 /// <reference path="./runtime-path.d.ts" />
 /// <reference path="./runtime-fs.d.ts" />
+/// <reference path="./runtime-net.d.ts" />